Check existence, not value of identity
In relations with type Backbone.One, _setAttr was checking that data's id
has a truthy value before comparing the identity with currVal's. This
means that if id's value is falsy, such as 0, the predicate fails even if
the actual values were the same. By checking for the id's existence
instead, we match things up correctly even with falsy identities.